+++ mom6/stats/mom6_regrid_tercile.py
